+++ src/edclasses/classes.py
@@ -4,61 +4,7 @@ from functools import cached_property
 from typing import Optional, List
 
 from . import enums
-
-_NEVER_EXPIRE = 0
-
-
-class ExpiringCachedPropertyMixin:
-    @staticmethod
-    def _get_expiration_key(item):
-        return f"{item}_expiration_time"
-
-    def _get_new_expiration_time(self, lifetime_in_seconds):
-        if lifetime_in_seconds:
-            return datetime.datetime.utcnow() + datetime.timedelta(
-                seconds=lifetime_in_seconds
-            )
-        else:
-            return None
-
-    @staticmethod
-    def _is_expired(expiration_time):
-        if expiration_time and datetime.datetime.utcnow() >= expiration_time:
-            return True
-        return False
-
-    def __getattribute__(self, item):
-        get_attr = super().__getattribute__
-        registry = get_attr("expiring_properties_registry")
-        try:
-            item_lifetime = registry[item]
-        except KeyError:
-            return get_attr(item)
-
-        cache = get_attr("__dict__")
-        expiration_key = get_attr("_get_expiration_key")(item)
-        expiration_time = cache.get(expiration_key)
-        time_expired = expiration_time and datetime.datetime.utcnow() >= expiration_time
-        if time_expired:
-            cache.pop(item, None)
-
-        val = super().__getattribute__(item)
-
-        if expiration_time is None or time_expired:
-            cache[expiration_key] = get_attr("_get_new_expiration_time")(
-                lifetime_in_seconds=item_lifetime
-            )
-        return val
-
-    def __setattr__(self, key, value):
-        get_attr = super().__getattribute__
-        registry = get_attr("expiring_properties_registry")
-        __dict__ = get_attr("__dict__")
-        if key in registry:
-            expiration_key = get_attr("_get_expiration_key")(key)
-            __dict__[expiration_key] = _NEVER_EXPIRE
-
-        return super().__setattr__(key, value)
+from .commons.caching_utils import ExpiringCachedPropertyMixin
 
 
 class System:
@@ -88,12 +34,14 @@ class FactionBranch(ExpiringCachedPropertyMixin):
         is_main: bool = False,
         influence: Decimal = 0,
         stations: List["OrbitalStation"] = None,
+        color="blue",
     ):
         self.faction = faction
         self.system = system
         self.is_main = is_main
         self.influence = influence
         self.stations = stations or []
+        self.color = "blue"
 
     def __str__(self):
         return f"{self.faction} in {self.system}"
